gamerle/requirements.txt
requests
Pillow
PyQt5